outsourced drivers.SAFE.removeGRDBorderNoise to S1.auxil function
the method drivers.SAFE.removeGRDBorderNoise still exists but now calls
S1.auxil.removeGRDBorderNoise in the background
