+++ pyroSAR/S1/auxil.py
@@ -14,6 +14,11 @@ import re
 import ssl
 import time
 from datetime import datetime
+import xml.etree.ElementTree as ET
+import numpy as np
+from osgeo import gdal
+from osgeo.gdalconst import GA_Update
+from . import linesimplify as ls
 
 from spatialist.ancillary import finder, urlQueryParser
 
@@ -395,3 +400,144 @@ class OSV(object):
             files_res = self.catch('RES', start=self.maxdate('RES', 'start'))
             self.retrieve(files_res)
             self.clean_res()
+
+
+def removeGRDBorderNoise(scene):
+    """
+    mask out Sentinel-1 image border noise
+
+    Reference:
+        Masking "No-value" Pixels on GRD Products generated by the Sentinel-1 ESA IPF' (issue 1, June 2015; issue 2.1 Jan 29 2018)
+        available online under https://sentinel.esa.int/web/sentinel/user-guides/sentinel-1-sar/document-library
+    
+    Parameters
+    ----------
+    scene: ~pyroSAR.drivers.SAFE
+        the Sentinel-1 scene object
+
+    Returns
+    -------
+
+    """
+    if scene.compression is not None:
+        raise RuntimeError('scene is not yet unpacked')
+
+    blocksize = 2000
+
+    # compute noise scaling factor
+    if scene.meta['IPF_version'] >= 2.9:
+        print('border noise removal not necessary for IPF version {}'.format(scene.meta['IPF_version']))
+        return
+    elif scene.meta['IPF_version'] <= 2.5:
+        knoise = {'IW': 75088.7, 'EW': 56065.87}[scene.acquisition_mode]
+        cads = scene.getFileObj(scene.findfiles('calibration-s1[ab]-[ie]w-grd-(?:hh|vv)')[0])
+        caltree = ET.fromstring(cads.read())
+        cads.close()
+        adn = float(caltree.find('.//calibrationVector/dn').text.split()[0])
+        if scene.meta['IPF_version'] < 2.34:
+            scalingFactor = knoise * adn
+        else:
+            scalingFactor = knoise * adn * adn
+    else:
+        scalingFactor = 1
+
+    # read noise vectors from corresponding annotation xml
+    noisefile = scene.getFileObj(scene.findfiles('noise-s1[ab]-[ie]w-grd-(?:hh|vv)')[0])
+    noisetree = ET.fromstring(noisefile.read())
+    noisefile.close()
+    noiseVectors = noisetree.findall('.//noiseVector')
+
+    # define boundaries of image subsets to be masked (4x the first lines/samples of the image boundaries)
+    subsets = [(0, 0, blocksize, scene.lines),
+               (0, 0, scene.samples, blocksize),
+               (scene.samples - blocksize, 0, scene.samples, scene.lines),
+               (0, scene.lines - blocksize, scene.samples, scene.lines)]
+
+    # extract column indices of noise vectors
+    yi = np.array([int(x.find('line').text) for x in noiseVectors])
+
+    # create links to the tif files for a master co-polarization and all other polarizations as slaves
+    master = scene.findfiles('s1.*(?:vv|hh).*tiff')[0]
+    ras_master = gdal.Open(master, GA_Update)
+    ras_slaves = [gdal.Open(x, GA_Update) for x in scene.findfiles('s1.*tiff') if x != master]
+
+    outband_master = ras_master.GetRasterBand(1)
+    outband_slaves = [x.GetRasterBand(1) for x in ras_slaves]
+
+    # iterate over the four image subsets
+    for subset in subsets:
+        print(subset)
+        xmin, ymin, xmax, ymax = subset
+        xdiff = xmax - xmin
+        ydiff = ymax - ymin
+        # linear interpolation of noise vectors to array
+        noise_interp = np.empty((ydiff, xdiff), dtype=float)
+        for i in range(0, len(noiseVectors)):
+            if ymin <= yi[i] <= ymax:
+                # extract row indices of noise vector
+                xi = map(int, noiseVectors[i].find('pixel').text.split())
+                # extract noise values
+                noise = map(float, noiseVectors[i].find('noiseLut').text.split())
+                # interpolate values along rows
+                noise_interp[yi[i] - ymin, :] = np.interp(range(0, xdiff), xi, noise)
+        for i in range(0, xdiff):
+            yi_t = yi[(ymin <= yi) & (yi <= ymax)] - ymin
+            # interpolate values along columns
+            noise_interp[:, i] = np.interp(range(0, ydiff), yi_t, noise_interp[:, i][yi_t])
+    
+        # read subset of image to array and subtract interpolated noise (denoising)
+        mat_master = outband_master.ReadAsArray(*[xmin, ymin, xdiff, ydiff])
+        denoisedBlock = mat_master.astype(float) ** 2 - noise_interp * scalingFactor
+        # mask out all pixels with a value below 0.5 in the denoised block or 30 in the original block
+        denoisedBlock[(denoisedBlock < 0.5) | (mat_master < 30)] = 0
+        denoisedBlock = np.sqrt(denoisedBlock)
+    
+        # mask out negative values
+        def helper1(x):
+            return len(x) - np.argmax(x > 0)
+    
+        def helper2(x):
+            return len(x) - np.argmax(x[::-1] > 0)
+    
+        if subset == (0, 0, blocksize, scene.lines):
+            border = np.apply_along_axis(helper1, 1, denoisedBlock)
+            border = blocksize - np.array(ls.reduce(border))
+            for j in range(0, ydiff):
+                denoisedBlock[j, :border[j]] = 0
+                denoisedBlock[j, border[j]:] = 1
+        elif subset == (0, scene.lines - blocksize, scene.samples, scene.lines):
+            border = np.apply_along_axis(helper2, 0, denoisedBlock)
+            border = ls.reduce(border)
+            for j in range(0, xdiff):
+                denoisedBlock[border[j]:, j] = 0
+                denoisedBlock[:border[j], j] = 1
+        elif subset == (scene.samples - blocksize, 0, scene.samples, scene.lines):
+            border = np.apply_along_axis(helper2, 1, denoisedBlock)
+            border = ls.reduce(border)
+            for j in range(0, ydiff):
+                denoisedBlock[j, border[j]:] = 0
+                denoisedBlock[j, :border[j]] = 1
+        elif subset == (0, 0, scene.samples, blocksize):
+            border = np.apply_along_axis(helper1, 0, denoisedBlock)
+            border = blocksize - np.array(ls.reduce(border))
+            for j in range(0, xdiff):
+                denoisedBlock[:border[j], j] = 0
+                denoisedBlock[border[j]:, j] = 1
+    
+        mat_master[denoisedBlock == 0] = 0
+        # write modified array back to original file
+        outband_master.WriteArray(mat_master, xmin, ymin)
+        outband_master.FlushCache()
+        # perform reading, masking and writing for all other polarizations
+        for outband in outband_slaves:
+            mat = outband.ReadAsArray(*[xmin, ymin, xdiff, ydiff])
+            mat[denoisedBlock == 0] = 0
+            outband.WriteArray(mat, xmin, ymin)
+            outband.FlushCache()
+    # detach file links
+    outband_master = None
+    ras_master = None
+    for outband in outband_slaves:
+        outband = None
+    for ras in ras_slaves:
+        ras = None
